requirements.txt
flask==3.1.0
gevent==24.11.1
# tensorflow==2.18.0
tensorflow==2.15.0
numpy==1.26.4
nltk==3.9.1
# gensim==4.3.3
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
PyMuPDF==1.24.14